+++ lib/preface_make/joker.mli
@@ -1,0 +1,20 @@
+(** [Joker] can produces [Bifunctor] or [Profunctor] using a [Functor] on the
+    second argument of the [Bi/Profunctor] as described in
+    {{:https://citeseerx.ist.psu.edu/viewdoc/download?doi=10.1.1.475.6134&rep=rep1&type=pdf}
+    Clowns to the Left, Jokers to the Right (Functional Pearl)}*)
+
+(** {2 Bifunctor}
+
+    Produces a {!module-type:Preface_specs.BIFUNCTOR} using a
+    {!module-type:Preface_specs.FUNCTOR} as second argument.*)
+
+module To_bifunctor (F : Preface_specs.Functor.CORE) :
+  Preface_specs.BIFUNCTOR with type ('a, 'b) t = 'b F.t
+
+(** {2 Profunctor}
+
+    Produces a {!module-type:Preface_specs.PROFUNCTOR} using a
+    {!module-type:Preface_specs.CONTRAVARIANT} as second argument.*)
+
+module To_profunctor (F : Preface_specs.Functor.CORE) :
+  Preface_specs.PROFUNCTOR with type ('a, 'b) t = 'b F.t