+++ lib/preface_make/preface_make.mli
@@ -92,3 +92,14 @@ module Free_applicative = Free_applicative
 module Free_selective = Free_selective
 module Free_monad = Free_monad
 module Freer_monad = Freer_monad
+
+(** {1 Conversion and Expansion}
+
+    Produces abstractions based on other abstractions (e.g. bifunctors from
+    functors). Mainly for dealing with different arities. *)
+
+module Join = Join
+module Clown = Clown
+module Joker = Joker
+module Kleisli = Kleisli
+module Cokleisli = Cokleisli